+++ src/modules/users/useCases/CreateUser/CreateUserUseCase.spec.ts
@@ -1,8 +1,8 @@
 import { v4 } from 'uuid';
-import { AppError } from '../../../shared/errors/AppError';
-import { User } from '../entities/User';
-import { FakeHashProvider } from '../providers/HashProvider/fakes/FakeHashProvider';
-import { FakeUsersRepository } from '../repositories/fakes/FakeUsersRepository';
+import { AppError } from '../../../../shared/errors/AppError';
+import { User } from '../../entities/User';
+import { FakeHashProvider } from '../../providers/HashProvider/fakes/FakeHashProvider';
+import { FakeUsersRepository } from '../../repositories/fakes/FakeUsersRepository';
 import { CreateUserUseCase } from './CreateUserUseCase';
 
 describe('CreateUserUseCase', () => {
