+++ src/modules/users/useCases/CreateUser/CreateUserUseCase.ts
@@ -1,9 +1,9 @@
 import { inject, injectable } from 'tsyringe';
 import { v4 } from 'uuid';
-import { AppError } from '../../../shared/errors/AppError';
-import { User } from '../entities/User';
-import { IHashProvider } from '../providers/HashProvider/models/IHashProvider';
-import { IUsersRepository } from '../repositories/IUsersRepository';
+import { AppError } from '../../../../shared/errors/AppError';
+import { User } from '../../entities/User';
+import { IHashProvider } from '../../providers/HashProvider/models/IHashProvider';
+import { IUsersRepository } from '../../repositories/IUsersRepository';
 import { CreateUserDTO } from './CreateUserDTO';
 
 @injectable()
